backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
SQLAlchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
orjson==3.10.7
tenacity==9.0.0
pytest==8.3.3
httpx==0.27.2
ruff==0.6.8
pyright==1.1.381
